Refactor : 게시글 수정, 삭제 - 게시글 수정, 삭제 권한 관련 코드 수정 (게시글 작성자만 수정/삭제 가능)
Related to: #4

+++ src/main/java/com/preonboarding/wanted/service/PostService.java
@@ -2,7 +2,6 @@ package com.preonboarding.wanted.service;
 
 import com.preonboarding.wanted.dto.request.SavePostRequest;
 import com.preonboarding.wanted.dto.request.UpdatePostRequest;
-import com.preonboarding.wanted.dto.response.DeletePostResponse;
 import com.preonboarding.wanted.dto.response.GetPostResponse;
 import com.preonboarding.wanted.dto.response.PagingPostResponse;
 import com.preonboarding.wanted.dto.response.SavePostResponse;
@@ -11,12 +10,15 @@ import com.preonboarding.wanted.entity.Post;
 import com.preonboarding.wanted.entity.User;
 import com.preonboarding.wanted.repository.PostRepository;
 import com.preonboarding.wanted.repository.UserRepository;
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
@@ -75,32 +77,47 @@ public class PostService {
     }
 
     @Transactional
-    public UpdatePostResponse updatePost(Long postId, UpdatePostRequest requestDto) {
+    public UpdatePostResponse updatePost(Long postId, UpdatePostRequest requestDto, Principal principal) {
 
-        Optional<Post> byId = postRepository.findById(postId);
-        Post post = byId.orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+        Optional<Post> findPost = postRepository.findById(postId);
 
-        post.setTitle(requestDto.getTitle());
-        post.setContent(requestDto.getContent());
+        if(findPost.get().getUser().getEmail().equals(principal.getName())) {
 
-        return UpdatePostResponse
-                .builder()
-                .result("게시글 수정이 완료되었습니다.")
-                .build();
+            Post post = findPost.orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+            post.setTitle(requestDto.getTitle());
+            post.setContent(requestDto.getContent());
+
+            return UpdatePostResponse
+                    .builder()
+                    .result("게시글 수정이 완료되었습니다.")
+                    .build();
+        } else  {
+            return UpdatePostResponse
+                    .builder()
+                    .result("게시글 수정 권한이 없습니다.")
+                    .build();
+        }
+
     }
 
     @Transactional
-    public DeletePostResponse deletePost(Long postId) {
+    public ResponseEntity<?> deletePost(Long postId, Principal principal) {
 
-        Optional<Post> byId = postRepository.findById(postId);
-        Post post = byId.orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+        Optional<Post> findPost = postRepository.findById(postId);
 
-        postRepository.delete(post);
+        if(findPost.get().getUser().getEmail().equals(principal.getName())) {
 
-        return DeletePostResponse
-                .builder()
-                .result("게시글 삭제가 완료되었습니다.")
-                .build();
+            System.out.println("findPost.get().getUser().getEmail() = " + findPost.get().getUser().getEmail());
+            System.out.println("principal.getName() = " + principal.getName());
+
+            Post post = findPost.orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+            postRepository.delete(post);
+
+            return ResponseEntity.ok("게시글이 정상적으로 삭제되었습니다.");
+        } else  {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("게시글 삭제 권한이 없습니다.");
+        }
+
     }
 
     @Transactional(readOnly = true)
